working sin object
