+++ code/trigo.py
@@ -2,19 +2,24 @@
 
 import math
 
+# from linear import x_simple as x_simple
+from linear import *
+
 class trigo():
 	'''
 	Toy forward automatic differentiation
 	class.
 	E.g.
-	f(x) = alpha * sin(gamma * x) + beta
+	f(x) = alpha * sin(x) + beta
+
+	Note:
+	x is a class object
 	'''
-	def __init__(self, a, alpha=1, beta=0, gamma=1):
+	def __init__(self, a, alpha=1, beta=0):
 
 		self.a = a # value to evaluate at
 		self.alpha = alpha # regard as a x variable with coefficient of x = 1
 		self.beta = beta # regard as a x variable with no constant
-		self.gamma = gamma # regard as the coefficient of the 
 
 		self.val = self.calc_function_val()
 		self.der = self.calc_function_derivative_val()
@@ -50,97 +55,106 @@ class trigo():
 		self.der = self.calc_function_derivative_val()
 
 
-	def __add__(self, other):
-		'''
-		Performs addition of two autodifftoyobjects,
-		or autodifftoyobject with a float/int
-		'''
+	# def __add__(self, other):
+	# 	'''
+	# 	Performs addition of two trigo objects,
+	# 	or autodifftoy object with a float/int
+	# 	'''
 
-		# Assume that both objects are AutoDiffToyObjects
-		try:
-			alpha = self.alpha + other.alpha
-			beta = self.beta + other.beta
-			new_toy = AutoDiffToy(self.a, alpha, beta)
-			return(new_toy)
+	# 	# Assume that both objects are AutoDiffToyObjects
+	# 	try:
+	# 		alpha = self.alpha + other.alpha
+	# 		beta = self.beta + other.beta
+	# 		new_toy = AutoDiffToy(self.a, alpha, beta)
+	# 		return(new_toy)
 
-		# Perhaps the 'other' is not an AutoDiffToyObject.
-		# So we'll just add the constant values
-		except:
-			try:
-				beta = self.beta + other.real
-				new_toy = AutoDiffToy(self.a, self.alpha, beta)
-				return(new_toy)
-			except:
-				raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
-
-
-	def __radd__(self, other):
-		'''
-		Allows for commuative cases of addition, where a
-		float or integer are added to the autodifftoy object.
-		'''
-		# Assume that both objects are AutoDiffToyObjects
-		try:
-			alpha = self.alpha + other.alpha
-			beta = self.beta + other.beta
-			new_toy = AutoDiffToy(self.a, alpha, beta)
-			return(new_toy)
-		# Perhaps the 'other' is not an AutoDiffToyObject.
-		# So we'll just add the constant values
-		except:
-			try:
-				beta = self.beta + other.real
-				new_toy = AutoDiffToy(self.a, self.alpha, beta)
-				return(new_toy)
-			except:
-				#pass
-				raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
+	# 	# Perhaps the 'other' is not an AutoDiffToyObject.
+	# 	# So we'll just add the constant values
+	# 	except:
+	# 		try:
+	# 			beta = self.beta + other.real
+	# 			new_toy = AutoDiffToy(self.a, self.alpha, beta)
+	# 			return(new_toy)
+	# 		except:
+	# 			raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
 
 
-	def __mul__(self, other):
-		'''
-		This allows for multiplication between a coefficient 
-		value and a AutoDiffToy object
-		'''
-		# Multiply a number with a 'x' class
-		try:
-			alpha = self.alpha * other.real
-			new_toy = AutoDiffToy(self.a, alpha, self.beta)
-			return(new_toy)
-
-		# Catch weird cases. E.g. when we're multiplying two 'x' classes
-		except:
-			raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
-
-
-	def __rmul__(self, other):
-		'''
-		We use rmul as it allows us to deal with commutative case
-		in which we have to multiply a float value with a AutoDiffToy
-		class, without having to overlaod __mul__ function in the
-		float class. (__mul__ function is called from the first variable
-		in a multiplication)
-		'''
-		try:
-			alpha = self.alpha * other.real
-			new_toy = AutoDiffToy(self.a, alpha, self.beta)
-			return(new_toy)
-		except:
-			raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
+	# def __radd__(self, other):
+	# 	'''
+	# 	Allows for commuative cases of addition, where a
+	# 	float or integer are added to the autodifftoy object.
+	# 	'''
+	# 	# Assume that both objects are AutoDiffToyObjects
+	# 	try:
+	# 		alpha = self.alpha + other.alpha
+	# 		beta = self.beta + other.beta
+	# 		new_toy = AutoDiffToy(self.a, alpha, beta)
+	# 		return(new_toy)
+	# 	# Perhaps the 'other' is not an AutoDiffToyObject.
+	# 	# So we'll just add the constant values
+	# 	except:
+	# 		try:
+	# 			beta = self.beta + other.real
+	# 			new_toy = AutoDiffToy(self.a, self.alpha, beta)
+	# 			return(new_toy)
+	# 		except:
+	# 			#pass
+	# 			raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
 
 
-class sin(trigo):
+	# def __mul__(self, other):
+	# 	'''
+	# 	This allows for multiplication between a coefficient 
+	# 	value and a AutoDiffToy object
+	# 	'''
+	# 	# Multiply a number with a 'x' class
+	# 	try:
+	# 		alpha = self.alpha * other.real
+	# 		new_toy = AutoDiffToy(self.a, alpha, self.beta)
+	# 		return(new_toy)
+
+	# 	# Catch weird cases. E.g. when we're multiplying two 'x' classes
+	# 	except:
+	# 		raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
+
+
+	# def __rmul__(self, other):
+	# 	'''
+	# 	We use rmul as it allows us to deal with commutative case
+	# 	in which we have to multiply a float value with a AutoDiffToy
+	# 	class, without having to overlaod __mul__ function in the
+	# 	float class. (__mul__ function is called from the first variable
+	# 	in a multiplication)
+	# 	'''
+	# 	try:
+	# 		alpha = self.alpha * other.real
+	# 		new_toy = AutoDiffToy(self.a, alpha, self.beta)
+	# 		return(new_toy)
+	# 	except:
+	# 		raise AttributeError(f'{other.__class__.__name__}.{name} is invalid for multiplication.')
+
+
+class sin():
 	'''
 	E.g.
-	f(x) = alpha * sin(gamma * x) + beta
+	f(x) = alpha * sin(x) + beta
 	'''
+	def __init__(self, x_object, alpha=1, beta=0):
+		#self.a = a # value to evaluate at
+		self.alpha = alpha # regard as a x variable with coefficient of sin
+		self.beta = beta # regard as a x variable with no constant
+		self.x_object = x_object # regard as a simple x_object class within the sine
+
+		self.val = self.calc_function_val()
+		self.der = self.calc_function_derivative_val()
+
 
 	def calc_function_val(self):
 		'''
 		Calculate the current value of this
 		function at the 
 		'''
-		value = self.alpha * math.sin(self.gamma * self.a) + self.beta
+		value = self.alpha * math.sin(self.x_object.val) + self.beta
 
 		return value
 
@@ -150,91 +164,99 @@ class sin(trigo):
 		Calculate the derivative of this function
 		at the a value of interest
 		'''
-
-		derivative_val = self.alpha * self.gamma * math.cos(self.gamma * self.a)
+		derivative_val = self.alpha * self.x_object.der * math.cos(self.x_object.val)
 
 		return derivative_val
 
 
-class cos(trigo):
-	'''
-	E.g.
-	f(x) = alpha * cos(gamma * x) + beta
-	'''
-
-	def calc_function_val(self):
+	def update_value_and_derivative(self):
 		'''
-		Calculate the current value of this
-		function at the 
+		Run functions to update the function values
+		and derivatives
 		'''
-		value = self.alpha * math.cos(self.gamma * self.a) + self.beta
+		self.val = self.calc_function_val()
+		self.der = self.calc_function_derivative_val()
 
-		return value
+
+# class cos(trigo):
+# 	'''
+# 	E.g.
+# 	f(x) = alpha * cos(gamma * x) + beta
+# 	'''
+
+# 	def calc_function_val(self):
+# 		'''
+# 		Calculate the current value of this
+# 		function at the 
+# 		'''
+# 		value = self.alpha * math.cos(self.a) + self.beta
+
+# 		return value
 
 	
-	def calc_function_derivative_val(self):
-		'''
-		Calculate the derivative of this function
-		at the a value of interest
-		'''
+# 	def calc_function_derivative_val(self):
+# 		'''
+# 		Calculate the derivative of this function
+# 		at the a value of interest
+# 		'''
 
-		derivative_val = (-1) * self.alpha * self.gamma * math.sin(self.gamma * self.a)
+# 		derivative_val = (-1) * self.alpha * math.sin(self.a)
 
-		return derivative_val
-
-
-class tan(trigo):
-	'''
-	E.g.
-	f(x) = alpha * tax(gamma * x) + beta
-	'''
-
-	def calc_function_val(self):
-		'''
-		Calculate the current value of this
-		function at the 
-		'''
-		value = self.alpha * math.cos(self.gamma * self.a) + self.beta
-
-		return value
-
-	
-	def calc_function_derivative_val(self):
-		'''
-		Calculate the derivative of this function
-		at the a value of interest
-		'''
-
-		derivative_val = self.alpha * self.gamma * (1 / math.cos(self.gamma * self.a) ** 2)
-
-		return derivative_val
+# 		return derivative_val
 
 
-class exponential(trigo):
-	'''
-	E.g.
-	f(x) = alpha * exp(gamma * x) + beta
-	'''
+# class tan(trigo):
+# 	'''
+# 	E.g.
+# 	f(x) = alpha * tax(gamma * x) + beta
+# 	'''
 
-	def calc_function_val(self):
-		'''
-		Calculate the current value of this
-		function at the 
-		'''
-		value = self.alpha * math.exp(self.gamma * self.a) + self.beta
+# 	def calc_function_val(self):
+# 		'''
+# 		Calculate the current value of this
+# 		function at the 
+# 		'''
+# 		value = self.alpha * math.cos(self.a) + self.beta
 
-		return value
+# 		return value
 
 	
-	def calc_function_derivative_val(self):
-		'''
-		Calculate the derivative of this function
-		at the a value of interest
-		'''
+# 	def calc_function_derivative_val(self):
+# 		'''
+# 		Calculate the derivative of this function
+# 		at the a value of interest
+# 		'''
 
-		derivative_val = self.alpha * self.gamma * math.exp(self.gamma * self.a)
+# 		derivative_val = self.alpha * self.gamma * (1 / math.cos(self.gamma * self.a) ** 2)
 
-		return derivative_val
+# 		return derivative_val
+
+
+# class exponential(trigo):
+# 	'''
+# 	E.g.
+# 	f(x) = alpha * exp(gamma * x) + beta
+# 	'''
+
+# 	def calc_function_val(self):
+# 		'''
+# 		Calculate the current value of this
+# 		function at the 
+# 		'''
+# 		value = self.alpha * math.exp(self.gamma * self.a) + self.beta
+
+# 		return value
+
+	
+# 	def calc_function_derivative_val(self):
+# 		'''
+# 		Calculate the derivative of this function
+# 		at the a value of interest
+# 		'''
+
+# 		derivative_val = self.alpha * self.gamma * math.exp(self.gamma * self.a)
+
+# 		return derivative_val
 
 
 
@@ -244,10 +266,27 @@ alpha = 2.0
 beta = 3.0
 gamma = 4.0
 
-x = sin(a=a, gamma=gamma)
-#f = alpha * x + beta
-f = x + beta
+
+
+a = 2.0 # Value to evaluate at
+x = AutoDiffToy(a)
+
+alpha = 2.0
+beta = 3.0
+f = alpha * x + beta
+
 print(f.val, f.der)
+
+
+x_2 = sin(f)
+print(x_2.val, x_2.der)
+
+
+
+# x = sin(a=a)
+# #f = alpha * x + beta
+# f = x + beta
+# print(f.val, f.der)
 
 
 #f = alpha * x + beta
